rename commands to lists
